requirements.txt
arrow==1.3.0
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
types-python-dateutil==2.9.0.20240906
urllib3==2.2.3
